requirements.txt
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
google-api-core==1.24.1
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-cloud-bigquery==2.6.1
google-cloud-bigquery-storage==2.1.0
google-cloud-core==1.5.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
idna==2.10
joblib==1.0.0
libcst==0.3.16
mypy-extensions==0.4.3
numpy==1.19.5
oauthlib==3.1.0
pandas==1.2.0
pandas-gbq==0.14.1
pkg-resources==0.0.0
proto-plus==1.13.0
protobuf==3.14.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydata-google-auth==1.1.0
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-learn==0.24.0
scipy==1.6.0
six==1.15.0
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
